[zsh] change location of including proxy environment variables
At least on macOS, putting them in zprofile (or even zlogin) and is
conceptually better than putting them in zshrc.  The problem is that
applications are ironically inheriting these proxy that shouldn't be, such as
Microsoft Teams and the macOS Software Updater, causing various issues with
them.  For this reason they shouldn't go in zshenv either.